eve7: rename GL/glu.h to GL_glu.h to avoid any conflicts with standard libs
